add error handling when the server crash on top-level expressions

+++ scripts/dev.js
@@ -118,7 +118,7 @@ function fix() {
   )
   fs.writeFileSync(file, text)
 }
-let stopServer = async () => {}
+let stopServer = () => Promise.resolve()
 let EPOCH = 0
 async function restartServer() {
   await stopServer()
@@ -131,10 +131,21 @@ async function restartServer() {
       EPOCH: String(EPOCH),
     },
   })
-  stopServer = () =>
-    new Promise(resolve => {
-      log('stopping server...')
-      server.on('close', resolve)
-      server.kill()
+  let stopped = false
+  let stopServerPromise = new Promise(resolve => {
+    server.on('close', () => {
+      stopped = true
+      log('server stopped')
+      resolve()
     })
+  })
+  stopServer = () => {
+    if (stopped) {
+      log('server already stopped')
+    } else {
+      log('stopping server...')
+      server.kill()
+    }
+    return stopServerPromise
+  }
 }
